% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read multiple years of FARS data}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector or list of integers
or characters of years for which to 
return the data}
}
\value{
This function returns a list of
  tibbles, one per specified year, with 
  one row per observation. The columns 
  return are MONTH and year.
}
\description{
This function returns a tibble of tidy
   data (one row per observation) with the
   month and year of the observation.
}
\details{
This function will return a warning
   if the data for a specified year 
   cannot be found.
   
Note: This function requires magrittr or tidyr
   is already loaded or it will return
   an error.
}
\examples{
\dontrun{
fars_read_years(2013)
fars_read_years(c(2013, 2014))
fars_read_years(2013:2015)
fars_read_years(c("2013", "2015"))
fars_read_years(list(2013, 2014))
}

}
