% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize FARS data across years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector or list of integers
or characters of years for which
to summarize the data}
}
\value{
This function returns a tibble
   with one row for each month and a 
   column for the month number followed
   by one column per specified year.
   These columns contain the observation
   counts for the relevant month and year.
}
\description{
This function creates a summary table
   of the counts of observations by 
   month across the specified years.
   
   This function will return a warning
   if the data for a specified year 
   cannot be found.
   
   Note: This function requires magrittr
   is already loaded or it will return
   an error.
}
\examples{
\dontrun{
fars_summarize_years(2013)   
fars_summarize_years(c(2013, 2014))
fars_summarize_years(2013:2015)
fars_summarize_years(c("2013", "2015"))
fars_summarize_years(list(2013, 2014))
}

}
